Thread: Fix stack overflow when using pthread_join

+++ OS_Common/Thread.cpp
@@ -15,7 +15,7 @@ int THREAD_JOIN(THREAD *thread)
 #ifdef WIN32
 	return WaitForSingleObject(*thread, INFINITE) == WAIT_FAILED ? -1 : 0;
 #else
-	int status;
+	long int status;
 	return pthread_join(*thread, (void **)&status);
 #endif
 }
